Move GhostLock Softlock Fix under Patches/Softlocks and tidy style.
Add SceneUtil for Scene 11, rename sit-tracking fields, wrap Update in try/catch, prefer var, and drop the oversized summary comment.
